requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
numpy==1.26.4
